check if the user deleting the post owns the post

+++ server/api/controllers/delete.post.js
@@ -2,15 +2,23 @@ const Post = require("../models/post.model");
 
 module.exports = (req, res) => {
   const postId = req.params.id;
-  Post.findByIdAndDelete(postId)
-    .exec()
-    .then((resp) => {
-      if (resp !== null) {
-        return res.json({ status: "success" });
+  Post.findById(postId)
+    .then((post) => {
+      if (post) {
+        if (post.posted_by == req.user.id)
+          return Post.findByIdAndDelete(postId)
+            .then(() => {
+              res.json({ status: "success" });
+            })
+            .catch((err) => {
+              res.json({ status: "error", message: err.message });
+            });
       }
-      return res.json({ status: "error", message: "post not found" });
+      return res
+        .status(404)
+        .json({ status: "error", message: "post not found" });
     })
-    .catch((err) => {
-      return res.status(500).json({ status: "error", message: err.message });
-    });
+    .catch((err) =>
+      res.status(500).json({ status: "error", message: err.message })
+    );
 };
